# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.createCommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
    type: object
...
  resp.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  resp.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SQL 审核、查询和工单管理平台 API
  title: SQLFlow API
  version: "1.0"
paths:
...
  /comments/{id}:
    delete:
      description: 删除指定的评论（仅作者或管理员可操作）
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 评论已删除
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的评论ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论
...
  /tickets/{id}/comments:
    get:
      description: 获取指定工单的所有评论
      parameters:
      - description: 工单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的工单ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 获取评论失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取工单评论列表
      tags:
      - 评论
    post:
      consumes:
      - application/json
      description: 在指定工单下创建评论
      parameters:
      - description: 工单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 工单不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"